requirements.txt
python-telegram-bot==20.7
langchain>=0.2.38
langchain-deepseek
#angchain-google-genai
#google-generativeai
langchain-core>=0.2.38
langchain-community
faiss-cpu>=1.13.0
python-dotenv
tqdm